Fix Next.js build errors: properly handle useSearchParams in client components, ensure Supabase URL has HTTPS prefix, and fix React components structure
